georeference onboarding, add
