+++ lib/assets/core/javascripts/cartodb3/components/onboardings/georeference/georeference-launcher.js
@@ -85,17 +85,19 @@ module.exports = (function () {
     },
 
     launch: function () {
-      var layerDefinitionModel = this._layerDefinitionsCollection.get(this._layerId);
+      if (!this._onboardingNotification.getKey(GEOREFERENCE_KEY)) {
+        var layerDefinitionModel = this._layerDefinitionsCollection.get(this._layerId);
 
-      AnalysesService.setLayerId(this._layerId);
+        AnalysesService.setLayerId(this._layerId);
 
-      this._onboardings.create(function () {
-        return new GeoreferenceOnboardingView({
-          onboardingNotification: this._onboardingNotification,
-          name: layerDefinitionModel.getName(),
-          notificationKey: GEOREFERENCE_KEY
-        });
-      }.bind(this));
+        this._onboardings.create(function () {
+          return new GeoreferenceOnboardingView({
+            onboardingNotification: this._onboardingNotification,
+            name: layerDefinitionModel.getName(),
+            notificationKey: GEOREFERENCE_KEY
+          });
+        }.bind(this));
+      }
     }
   };
 })();
