+++ lib/assets/core/javascripts/cartodb3/deep-insights-integrations.js
@@ -59,7 +59,6 @@ var F = function (opts) {
   this._layerDefinitionsCollection.each(this._linkLayerErrors, this);
 
   this._analysisDefinitionNodesCollection.on('add', this._analyseDefinitionNode, this);
-  this._analysisDefinitionNodesCollection.on('add', _.debounce(this._onAddAnalyseDefinitionNode, 120), this);
   this._analysisDefinitionNodesCollection.on('change', this._analyseDefinitionNode, this);
   this._analysisDefinitionNodesCollection.on('change:id', this._onAnalysisDefinitionNodeIdChanged, this);
   this._analysisDefinitionNodesCollection.on('remove', this._onAnalysisDefinitionNodeRemoved, this);
@@ -278,30 +277,6 @@ F.prototype._analyseDefinitionNode = function (m) {
 
     // Unfortunately have to try to setup sync until this point, since a node doesn't exist until after analyse call
     this._analysisDefinitionNodesCollection.each(this._tryToSetupDefinitionNodeSync, this);
-  }
-};
-
-F.prototype._onAddAnalyseDefinitionNode = function (m) {
-  var layerDefModel = this._layerDefinitionsCollection.findOwnerOfAnalysisNode(m);
-  var queryRowsCollection = m.queryRowsCollection;
-
-  var launchGeoreferenceOnboarding = function () {
-    if (!layerDefModel.hasAnalyses() &&
-        !layerDefModel.isCustomQueryApplied() &&
-        !queryRowsCollection.isEmpty()) {
-      GeoreferenceOnboardingLauncher.setLayerId(layerDefModel.get('id'));
-      GeoreferenceOnboardingLauncher.launch();
-    }
-  };
-
-  if (queryRowsCollection.shouldFetch()) {
-    queryRowsCollection.fetch({
-      success: launchGeoreferenceOnboarding
-    });
-  }
-
-  if (queryRowsCollection.isFetched()) {
-    launchGeoreferenceOnboarding();
   }
 };
