+++ lib/assets/core/javascripts/cartodb3/editor/editor-view.js
@@ -19,6 +19,7 @@ var Infobox = require('../components/infobox/infobox-factory');
 var InfoboxModel = require('../components/infobox/infobox-model');
 var InfoboxCollection = require('../components/infobox/infobox-collection');
 var AnalysesService = require('./layers/layer-content-views/analyses/analyses-service.js');
+var GeoreferenceOnboardingLauncher = require('../components/onboardings/georeference/georeference-launcher');
 
 var GEOREFERENCE_KEY = 'georeference';
 
@@ -54,9 +55,7 @@ module.exports = CoreView.extend({
     this._initModels();
     this._initBinds();
 
-    if (this._isGeoreferenceOnboardingEnabled()) {
-      this._findNonGeoreferencedLayer();
-    }
+    this._findNonGeoreferencedLayer();
   },
 
   render: function () {
@@ -226,14 +225,39 @@ module.exports = CoreView.extend({
     this.listenTo(this._widgetDefinitionsCollection, 'reset remove add', this._updateAddButtonState, this);
     this.listenTo(this._editorModel, 'change:edition', this._changeStyle, this);
     this.listenTo(this._layerDefinitionsCollection, 'reset remove add', function () {
-      if (this._isGeoreferenceOnboardingEnabled()) {
-        this._nonGeoreferencedLayer = null;
-        this._infoboxModel.unset('state');
+      this._nonGeoreferencedLayer = null;
+      this._infoboxModel.unset('state');
 
-        this._findNonGeoreferencedLayer();
-      }
+      this._findNonGeoreferencedLayer();
 
       this._updateAddButtonState();
+    }, this);
+    this.listenTo(this._layerDefinitionsCollection, 'add', function (model) {
+      var analysisDefinitionNodeModel = model.getAnalysisDefinitionNodeModel();
+      var queryRowsCollection = analysisDefinitionNodeModel.queryRowsCollection;
+
+      var launchGeoreferenceOnboarding = function () {
+        if (!model.hasAnalyses() &&
+            !model.isCustomQueryApplied() &&
+            !queryRowsCollection.isEmpty()) {
+          GeoreferenceOnboardingLauncher.setLayerId(model.get('id'));
+          GeoreferenceOnboardingLauncher.launch();
+        }
+      };
+
+      this.listenTo(queryRowsCollection.statusModel, 'change:status', function () {
+        if (queryRowsCollection.isFetched()) {
+          launchGeoreferenceOnboarding();
+        }
+      }, this);
+
+      if (queryRowsCollection.shouldFetch()) {
+        queryRowsCollection.fetch();
+      }
+
+      if (queryRowsCollection.isFetched()) {
+        launchGeoreferenceOnboarding();
+      }
     }, this);
     this.listenTo(this._layerDefinitionsCollection, 'add', function () {
       this._visDefinitionModel.fetch();
@@ -296,10 +320,6 @@ module.exports = CoreView.extend({
 
     AnalysesService.setLayerId(layerId);
     AnalysesService.addGeoreferenceAnalysis();
-  },
-
-  _isGeoreferenceOnboardingEnabled: function () {
-    return !this._onboardingNotification.getKey(GEOREFERENCE_KEY);
   },
 
   _onRemoveMap: function () {
